Moving motion interpolator to byte arrays too

+++ src/main/java/org/jcodec/codecs/h264/encode/MBEncoderP16x16.java
@@ -140,7 +140,7 @@ public class MBEncoderP16x16 {
     }
 
     private int[] mvEstimate(Picture8Bit pic, int mbX, int mbY, int mvpx, int mvpy) {
-        int[] patch = new int[256];
+        byte[] patch = new byte[256];
         MBEncoderHelper.takeSafe(pic.getPlaneData(0), pic.getPlaneWidth(0), pic.getPlaneHeight(0), mbX << 4, mbY << 4,
                 patch, 16, 16);
         return me.estimate(ref, patch, mbX, mbY, mvpx, mvpy);
